+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - Expected report count assertion
 - Force update snapshots when running `testRefApp`
+- Hint how to customize the AppImpactTest.
 
 ## [0.0.3]
 [0.0.3]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-0.0.3%0Drelease-0.0.2